refactor: Update rpcHandler to include host in DecodeRequestBody
The rpcHandler function in handler.go has been updated to include the host parameter in the DecodeRequestBody function call. This change ensures that the set1weiGasPrice function is called with the correct host and method parameters, allowing it to set the gas price to "0x1" for requests to the "eth_gasPrice" method on the "0.48.club" host.